Add sort-exports eslint plugin
This will allow us to ensure exports are sorted in particular orders in
certain cases.  This won't be applied to the whole project, only for our
type definitions which will follow a model of having an `index.js` that
exports each individual type.  We want that to be in alphabetical order
to prevent situations where it's not clear which files are / are not
being exported.