Add ArchiveVo type definitions
These types are all related to the ArchiveVo model design, which is
returned by the Permanent PHP API.

The ArchiveVO reflects the raw information that is sent to / from the
Permanent PHP API.  These fields were taken from the objects defined in
the back-end repository, specifically at `/library/base/vo`.

The ArchiveVo type relies on several other types which will need to be
populated in more detail before implementation can occur.  I'm choosing
to do that in a future commit to avoid having this become too large.

Several of these fields will not ultimately map to our exposed Archive
type / anything in the SDK, but this makes that lack of translation
explicit for future developers and ensures we're prepared to receive
whatever the server is sending (and send whatever the server is
expecting).

Issue #4