Add archiveVo default values
By providing a set of default values it becomes easy to create new
ArchiveVos without having to populate every single value.  This was
originally implemented as a factory method but it became clear that all
the factory method was doing was combining the default with the input --
no need for a function to do this.