Add enum typeguard generation utility
Enums are special in TypeScript in that they are converted to actual
usable runtime objects.

It will be a common use case in this project to need to take a string
(e.g. an API return value) and type it to one of our many enum values.
This utility function will prevent us from having to repeat the same
"keyof" / object meta logic for every single one of our typeguards.