Add "unknown" statuses
We are going to need to translate status strings provided to us by the
Permanent API into an enum that contains the complete set of valid
statuses.

What happens if the server provides an unknown value?

This allows us to specify that the value is unknown.  The alternative
would be to throw an error.