Split children into folders and items
The initial Archive type design had Folders and Items combined into
a single `children` attribute.  While it is true they are both children,
combining them would force our users to type guard each item to
determine if it is a child or folder.

It's very easy to combine the two child types if desired.

Issue #4