Change build logic
We want to build in a way that supports tree shaking in addition to node
use; this changes our build to create two versions of the code base (one
for classic js and the other for es modules).

While making this change I also changed our build directory to `dist` to
make it clear that the build is for distribution.

I also caught an issue where our configs were ignoring the wrong kind of
test file (.spec.js instead of .test.js).  When I fixed that issue it
uncovered a problem where our linter gets angry if there are any
matching files not included in the project.  All this is to say: there
is now a lint focused tsconfig.

Issue #20

We were building to a `lib` directory before