Downgrade node-fetch to v2
node-fetch@3 does not support Common JS (CJS) but we want our sdk to
support CJS at this point.  This means we can't use v3 and have to use
v2.

At this point the `node-fetch` project claims they will continue to
patch v2 so this is not a liability, but it may be a sign that the
future of node is modules / CJS will eventually be something we want to
stop supporting too.

Issue #25 node-fetch v3+ doesn't support common JS