Fix export order rule
We only care about export order for index files, which we use to export
a series of files for quality of life purposes.  That's where order is
important.

In non-index situations we do NOT want to be prescriptive on export
order since we might have exports that rely on previous exports
(regardless of name)