Add ajv
Ajv is a popular implementation of json schema and it provides some nice
runtime companionship to TypeScript.  In particular you can define
a schema in relation to a TypeScript type which can then be used as
a type guard.

Ajv is smart enough to yell if the schema and the type fall out of sync.

This also requires "strictNullChecks" to be enabled because otherwise
JSONSchemaType gets very angry.