Add permanent API call utility
This creates the ability to query against permanent in addition to
configuring the client that is making that call.

The ClientConfiguration will be passed to our sdk calls by end users and
reflect their "client".

It provides the authentication information the SDK needs as well as
information about where the permanent API instance they want to work
against is hosted.  This allows the SDK to be used in development
environments.

This commialso adds the nock package.

We need to be able to unit test parts of the sdk that rely on http
responses.  `nock` is a library that will intercept http calls in
a testing environment and allow our tests to mock a server response [1].

[1] https://www.npmjs.com/package/nock