Add utility for UTC formatting
The permanent backend stores all dates as UTC but does not actually
communicate them as UTC dates (it doesn't add `Z` to the end of the
date).

We never expect the UTC status to change, though it is possible that the
response will some day be corrected to include `Z`. In order to handle
both the current and future case we'll need to explicitly add the Z if
it doesn't exist so that dates are interpreted correctly.