Test package exports
These tests just make sure we're properly exporting the pieces of the
sdk that we want exported.  We might some day want to add additional
tests to ensure we're NOT exporting things that we don't want exposed
(and to explicitly list the methods we want exposed), but this at least
makes sure that index.ts works.