Remove unused types
The FolderSizeVo type was used in a previous iteration of FolderVo but
isn't used by the current code base; taking it out just to keep things
free of bloat (it may be added some day).

Item and ItemVo was also taken out of the picture.  This is beacause
Items are either Records or Folders and so instead of having an Item
/ ItemVo type we should just have any "Item" field be a union type.