Add getFolder sdk function
This provides the ability for an SDK user to access a folder object from
Permanent.  Permanent's permission system is a bit confusing: folders
cannot be accessed outside of the context of an archive.  This is not
something we are going to be able to change for the SDK, but it is the
reason why folder lookup also requires an archive ID.

Issue #43 Support folder access