Add `fileSystemCompatibleName` attributes
The `fileSystemCompatibleName` is a version of a given record or folder's name
that will work correctly on a file system.  This means it will not have
illegal characters and will not conflict with peer items.

This is a breaking change for any client that was using the
`Record.fileName` attribute.

Issue #62