Rename Record to ArchiveRecord
`Record` is already a concept in TypeScript [1]. This means that users
of our SDK would need to rename `Record` to something else if they
wanted to be able to still have access to TypeScript's `Record` utility
type.

We talked about a few options and landed on `ArchiveRecord`. This is
an existing term in the world of archival, and it also doesn't already
exist in the backend / server ("Item" was on the table too but the
permanent API already has the concept of an Item so it would have been
confusing to use that.)

The API itself is not changing and `Record` will continue to be a back-end
concept.

Issue #47

[1] https://www.typescriptlang.org/docs/handbook/utility-types.html#recordkeys-type