Add multipart form test utility
Multipart forms are a huge pain to test against and for whatever reason
it looks like nock doesn't support multipart body parsing out of the
box.  This little utility method should, in theory, be able to compare
a multipart form body with a list of key / value pairs and at least
confirm that each pair exists in the body somewhere.

This will only be used for unit tests.