Handle null values explicitly
There is a little bit of complexity with the way ajv currently handles
optional attributes in types.  In particular, ajv requires that those
fields be `nullable`.  In TypeScript the `?` flag just means a field
could be `undefined`, which is not the same as `nullable`.

This tweaks our types to adhere to ajv's current limitation, and then
also updates our code and tests to correct a bug that would occur when
DisplayDT was null.

Issue #75