Add slug to Archive type
Permanent's backend has the concept of an `archiveNbr` which is
equivalent to the MDN glossary's definition of a `slug`.  It serves as
the unique url based identifier of the archive.

This is a more user-centric identifier compared to the internal `id`,
and may be useful for client applications to show to the user in certain
contexts.

Issue #87