Throw on non-ok server responses
When the server responds with a non-ok status we want to communicate
that response to the calling code base via a thrown error.  Prior to
this commit server errors would often be completely ignored and could
result in odd validation errors instead of communicating the actual
issue.

As part of this change our delete functions now *throw* when something
goes wrong, and so the type signature has been updated to no longer
return `boolean` to indicate success but rather a `void`.

Issue #59 Handle server errors explicitly
Issue #89 Evalauate return types for operation methods