Remove commonjs specification
The new tsconfig for node18 requires `module` to be a valid value for
the specified `moduleResolution`.  Our override resulted in a conflict.

The `Node16` module setting *allows* esm exports.  This is a bit of
a moot point for that particular config since the whole point is that it
is a configuration for common JS.  Nevertheless, it feels better to not
override if we don't need to override, and the default behavior still
exports CJS.