Add fileSystemId field to Folders and ArchiveRecords
We are in the process of adding share link creation to this SDK. The
/share/generateShareLink endpoint takes a folder_linkId to determine
what it creates the share link for. folder_links are an overly
complicated construct that should be abstracted away in this SDK, but
they do capture the useful and important concept that records and
folders are in some ways the same kind of object. In other parts of
Permanent's code, this concept is captured in the notion of an Item, but
here "item" is already used to refer to a file that is not (yet) a
Permanent File object. In order to make folder_linkIds
available to users of this SDK for use with the incoming createShareLink
function, this commit adds fileSystemIds to the Folder and ArchiveRecord
interfaces, and populates them with the folder_linkIds of the
corresponding FolderVo and RecordVo interfaces.