Add updateShareLink function
We want to expose our /share/updateShareLink endpoint through this
SDK. To that end, this commit adds a new updateShareLink function that
hits that endpoint. It also makes several more fields of the ShareLink
interface read-only, to clarify which fields are user-configurable, and
adds documentation for the SDK, currently covering just the
share-related functions.