Use parameter interfaces for exposed functions
Some of the functions exposed by this SDK have a number of arguments
that strains readability. For increased readability, and to facilitate
the possible future addition of optional parameters to these functions,
this commit groups the parameters for each exposed function into a
params interface for that function. Each function now takes a
ClientConfiguration and a params interface specific to that function
(except a couple that just take the ClientConfiguration).