Use stela to load folders
The v2 folder load API endpoint will time out under certain
circumstances (e.g. for folders with a large number of children).  By
using Stela endpoints we are able to more reliably load folders.

This is the first time the sdk is interacting with Stela, which means
that we need to take the Stela API url in addition to the php base url.

Issue #479 Support new folder-children pagination endpoints