Add failOnDuplicateName flag to createFolder
We recently added a failOnDuplicateName flag to our /folder/post
endpoint to support the correct SFTP behavior of returning an error if
it tries to create a folder whose name is already in use at that path.
This commit updates the SDK's createFolder function to offer that
parameter. It is optional and defaults to false.