Enable eslint-love rules
We're using this package in our other projects as a solid and
opinionated lint rule set.

Since some of these rules are new we violate them.  Rather than making
a ton of fixes at once, we disable those rules for now and we can
re-enable them one by one.

Issue #541 Move to ESLint 9